Move ServerSoftware to top-level package
